pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dailyclaw"
version = "1.0.0"
description = "Plugin-based Telegram bot for daily journaling and accountability"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0",
    "python-telegram-bot[job-queue]>=21.0",
    "openai>=1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "jinja2>=3.1",
    "tzdata>=2024.1",
    "httpx[socks]>=0.27",
    "readability-lxml>=0.8",
    "lxml_html_clean>=0.4",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.35",
]

[project.scripts]
dailyclaw = "src.main:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.setuptools.package-data]
"src" = ["templates/*.html"]
"src.core" = ["migrations/*.sql"]
"src.plugins.recorder" = ["migrations/*.sql"]
"src.plugins.journal" = ["migrations/*.sql"]
"src.plugins.planner" = ["migrations/*.sql"]
"src.plugins.sharing" = ["migrations/*.sql"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
